split indexes components
